+++ codec/hessian/response.go
@@ -19,6 +19,10 @@ import (
 	jerrors "github.com/juju/errors"
 )
 
+import (
+	"github.com/AlexStocks/dubbogo/codec"
+)
+
 const (
 	Response_OK                byte = 20
 	Response_CLIENT_TIMEOUT    byte = 30
@@ -35,63 +39,83 @@ const (
 	RESPONSE_NULL_VALUE     int32 = 2
 )
 
-var (
-	ErrIllegalPackage = jerrors.Errorf("illegal pacakge!")
-)
-
 // hessian decode respone
-func UnpackResponse(buf []byte) (interface{}, error) {
+func unpackResponseHeaer(buf []byte, m *codec.Message) error {
 	length := len(buf)
-	if length < HEADER_LENGTH || (buf[0] != byte(MAGIC_HIGH) && buf[1] != byte(MAGIC_LOW)) {
-		return nil, ErrIllegalPackage
+	// hessianCodec.ReadHeader has check the header length
+	//if length < HEADER_LENGTH {
+	//	return codec.ErrHeaderNotEnough
+	//}
+
+	if buf[0] != byte(MAGIC_HIGH) && buf[1] != byte(MAGIC_LOW) {
+		return codec.ErrIllegalPackage
 	}
 
 	// Header{serialization id(5 bit), event, two way, req/response}
-	var serialID = buf[2] & SERIALIZATION_MASK
+	serialID := buf[2] & SERIALIZATION_MASK
 	if serialID == byte(0x00) {
-		return nil, jerrors.Errorf("serialization ID:%v", serialID)
+		return jerrors.Errorf("serialization ID:%v", serialID)
 	}
-	//var eventFlag byte = buf[2] & FLAG_EVENT
-	//if eventFlag == byte(0x00) {
-	//	return nil, jerrors.Errorf("event flag:%v", eventFlag)
-	//}
-	//var twoWayFlag byte = buf[2] & FLAG_TWOWAY
-	//if twoWayFlag == byte(0x00) {
-	//	return nil, jerrors.Errorf("twoway flag:%v", twoWayFlag)
-	//}
-	var rspFlag = buf[2] & FLAG_REQUEST
-	if rspFlag != byte(0x00) {
-		return nil, jerrors.Errorf("response flag:%v", rspFlag)
+
+	flag := buf[2] & FLAG_EVENT
+	if flag != byte(0x00) {
+		m.Type |= codec.Heartbeat
+	}
+	flag = buf[2] & FLAG_TWOWAY
+	if flag != byte(0x00) {
+		m.Type |= codec.Response
+	}
+	flag = buf[2] & FLAG_REQUEST
+	if flag != byte(0x00) {
+		return jerrors.Errorf("response flag:%v", flag)
 	}
 
 	// Header{status}
 	if buf[3] != Response_OK {
-		return nil, jerrors.Errorf("Response not OK, java exception:%s", string(buf[18:length-1]))
+		return jerrors.Errorf("Response not OK, java exception:%s", string(buf[18:length-1]))
 	}
 
 	// Header{req id}
-	//var ID int64 = int64(binary.BigEndian.Uint64(buf[4:]))
-	//fmt.Printf("response package id:%#X\n", ID)
+	m.ID = int64(binary.BigEndian.Uint64(buf[4:]))
 
 	// Header{body len}
-	var bodyLen = int32(binary.BigEndian.Uint32(buf[12:]))
-	if int(bodyLen+HEADER_LENGTH) != length {
-		return nil, ErrIllegalPackage
+	m.BodyLen = int(binary.BigEndian.Uint32(buf[12:]))
+	if m.BodyLen < 0 {
+		return codec.ErrIllegalPackage
 	}
 
+	return nil
+}
+
+// hessian decode response body
+func unpackResponseBody(buf []byte, ret interface{}) error {
 	// body
-	decoder := NewDecoder(buf[16:length])
-	rspObj, _ := decoder.Decode()
-	switch rspObj {
-	case RESPONSE_WITH_EXCEPTION:
-		return decoder.Decode()
-	case RESPONSE_VALUE:
-		return decoder.Decode()
-	case RESPONSE_NULL_VALUE:
-		return nil, jerrors.New("Received null")
+	decoder := NewDecoder(buf[:])
+	rspType, err := decoder.Decode()
+	if err != nil {
+		return jerrors.Trace(err)
 	}
 
-	return nil, nil
+	switch rspType {
+	case RESPONSE_WITH_EXCEPTION:
+		expt, err := decoder.Decode()
+		if err != nil {
+			return jerrors.Trace(err)
+		}
+		return jerrors.Errorf("got exception: %+v", expt)
+
+	case RESPONSE_VALUE:
+		rsp, err := decoder.Decode()
+		if err != nil {
+			return jerrors.Trace(err)
+		}
+		return jerrors.Trace(ReflectResponse(rsp, ret))
+
+	case RESPONSE_NULL_VALUE:
+		return jerrors.New("Received null")
+	}
+
+	return nil
 }
 
 func cpSlice(in, out interface{}) error {