+++ change_log.md
@@ -10,6 +10,11 @@ dubbogo 目前版本(0.1.1)支持的codec 是jsonrpc 2.0，transport protocol是
 ## develop list ##
 ---
 
+### 2018-05-17
+---
+- 1 把github.com/AlexStocks/gohessian最新的hessian2解析代码合并到dubbogo/codec/hessian下面；
+- 2 改进tcp transport socket字节流读取机制；
+
 ### 2018-05-07
 ---
 - 1 fmt.Errorf -> juju/errors.Errorf